move streaming read/write support interface to sources.v2 package
